% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_ottawa.R
\name{get_open_ottawa}
\alias{get_open_ottawa}
\title{Get Open Ottawa Data}
\usage{
get_open_ottawa(
  url_end,
  file_dest = paste0(getwd(), "/new_data.xlsx"),
  load_data = TRUE
)
}
\arguments{
\item{url_end}{the part of the url after https://open.ottawa.ca/datasets/}

\item{file_dest}{the location and name of the file to save make sure it ends with .xlsx}

\item{load_data}{should the data be loaded into a data table}
}
\description{
Function used to download single xlsx files from open ottawa
}
\examples{
tmp <- get_open_ottawa(url_end = "covid-19-hospitalization-data-for-ottawa-residents", file_dest = paste0(getwd(), "/new_data.xlsx"))

 @export
}
